CA-183332: Don't close stdin when listening for docker events
Calling stdin.channel.shutdown_write() worked fine in earlier environment
versions.
However, with the current versions in combination with certain guests
shutdown_write triggers the event call to return instantaneously (with rc=0),
instead of waiting and listening as we'd like it to. This is presumably caused
by a slightly different (i.e. more complete) implementation of shutdown_write.

Signed-off-by: Robert Breker <robert.breker@citrix.com>