Update util/__init__.py
Modified `convert_dict_to_ascii(item)` to convert to utf-8 instead. Some docker containers have values that are outside of the ascii range which causes xscontainer to crash with a `UnicodeEncodeError`. If ascii is absolutely necessary, line 154 could be changed to `item.encode('ascii', 'replace')` instead. I've tested the change to utf-8 on my system with no ill effects. 

If this is changed to utf-8, the function name and references should probably be updated to `convert_dict_to_utf8` as well. Didn't want to get ahead of myself in case this is a show stopper.